fix(edge-function): relocate eval functions to client side
