+++ apps/web/src/app/[lang]/app/[type]/crm/[domain]/[data]/(alternative)/details/[id]/form.tsx
@@ -27,9 +27,10 @@ import { useState } from "react";
 import { getResourceDataClient } from "src/language-data/CRMService";
 import { useLocale } from "src/providers/locale";
 import type { TableData } from "src/utils";
-import { getBaseLink, isPhoneValid, splitPhone } from "src/utils";
+import { getBaseLink} from "src/utils";
 import { dataConfigOfCrm } from "../../../../../data";
 import { updateCRMDetailServer, updateMerchantCRMDetailServer } from "./action";
+import { isPhoneValid, splitPhone } from "src/utilsPhone";
 
 const organization = $UniRefund_CRMService_Organizations_UpdateOrganizationDto;
 const email = $UniRefund_CRMService_EmailCommonDatas_UpdateEmailCommonDataDto;
