+++ apps/web/src/app/[lang]/app/[type]/crm/[domain]/[data]/(alternative)/new/page.tsx
@@ -11,7 +11,8 @@ import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 import type { TableData } from "src/utils";
-import { getBaseLink, isPhoneValid, splitPhone } from "src/utils";
+import { getBaseLink} from "src/utils";
+import { isPhoneValid, splitPhone } from "src/utilsPhone";
 import { getResourceDataClient } from "src/language-data/CRMService";
 import { useLocale } from "src/providers/locale";
 import { dataConfigOfCrm } from "../../../../data";
