+++ apps/web/src/app/[lang]/app/[type]/profile/[profileId]/form.tsx
@@ -9,7 +9,7 @@ import AutoForm, {
 } from "@repo/ayasofyazilim-ui/organisms/auto-form";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
-import { isPhoneValid, splitPhone } from "src/utils";
+import { isPhoneValid, splitPhone } from "src/utilsPhone";
 import {
   deleteBacker,
   postBacker,
